Altered seeduserWithRole() method name to match C# conventions.
Added method to UserService AddRoleToUser() to assign a specific role to a specific user. To do so, a RoleRepository class has been implemented with respective additions to the UnitOfWork class.

Added new functionality to ~/api/account/register, so that new registrated user are now automatically assigned to the customer-role.

This closes #40